a1/requirements.txt
# Core dependencies
aiohttp>=3.9.0
asyncio>=3.4.3
web3>=6.11.0
eth-account>=0.10.0
eth-typing>=3.5.0
hexbytes>=0.3.0

# Grok/LLM integration
openai>=1.3.0  # OpenRouter uses OpenAI-compatible API
tenacity>=8.2.0
tiktoken>=0.5.0

# Ethereum tools
py-solc-x>=2.0.0
eth-abi>=4.2.0
eth-utils>=2.3.0
multicall>=0.7.0

# Static analysis
slither-analyzer>=0.10.0
crytic-compile>=0.3.0

# Database and storage
asyncpg>=0.29.0
pgvector>=0.2.0
redis>=5.0.0
sqlalchemy>=2.0.0

# Agent framework
praison>=0.0.50
langchain>=0.1.0
langchain-community>=0.0.10

# Utilities
pyyaml>=6.0
python-dotenv>=1.0.0
click>=8.1.0
rich>=13.0.0
tabulate>=0.9.0

# Development tools
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
mypy>=1.5.0
ruff>=0.1.0

# Monitoring and logging
structlog>=23.0.0
prometheus-client>=0.19.0

# Security
cryptography>=41.0.0
pycryptodome>=3.19.0